抽出RxEditorAntdInner
